Add build script for compaction Modify compaction src for Hadoop 1
