--- compaction-tool/compaction.properties
@@ -1,3 +1,0 @@
-compaction.config.dir=jobconf
-hdfs.uri=
-hivesite.dir=
