--- compaction-tool/src/main/java/compactor/Compactor.java
@@ -1,29 +1,0 @@
-/* (c) 2015 LinkedIn Corp. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use
- * this file except in compliance with the License. You may obtain a copy of the
- * License at  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed
- * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
- * CONDITIONS OF ANY KIND, either express or implied.
- */
-
-package compactor;
-
-import java.io.IOException;
-
-
-/**
- * Compactor interface.
- *
- * @author ziliu
- */
-public interface Compactor {
-
-  /**
-   * A method for data compaction, e.g., merge a snapshot table and a delta table.
-   * @throws IOException 
-   */
-  public void compact() throws IOException;
-}
