+++ conf/log4j-compaction.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE log4j:configuration SYSTEM "log4j.dtd">
+
+<log4j:configuration xmlns:log4j="http://jakarta.apache.org/log4j/">
+  <appender name="console" class="org.apache.log4j.ConsoleAppender">
+    <param name="Target" value="System.out"/>
+    <layout class="org.apache.log4j.PatternLayout">
+      <param name="ConversionPattern" value="%d{HH:mm:ss} %5p %c{1} - %m%n"/>
+    </layout>
+  </appender>
+
+  <root>
+    <level value ="info" />
+    <appender-ref ref="console" />
+  </root>
+
+</log4j:configuration>
