+++ gobblin-compaction/compaction.properties
@@ -1,0 +1,3 @@
+compaction.config.dir=jobconf
+hdfs.uri=
+hivesite.dir=
