+++ gobblin-compaction/jobconf/task1.conf
@@ -1,0 +1,18 @@
+snapshot.name=
+snapshot.pkey=
+snapshot.datalocation=
+snapshot.copydata=
+snapshot.dataformat=
+
+delta.1.name=
+delta.1.pkey=
+delta.1.datalocation=
+delta.1.copydata=
+delta.1.dataformat=
+
+output.name=
+output.datalocation=
+
+hive.db.name=
+hive.tmpschema.dir=
+hive.tmpdata.dir=
